projeto_PI/codigo_paginas_formulario_e_DB/requirements.txt
streamlit
pydantic
sqlalchemy
python-dotenv
pymysql 